feat(wordpress): upgrade version, update plugins and themes (#18)
* .gitignore + rework .env

* Reworked docker system

* Fixed db exporter

* Updated README.md

* fix:(user.php):fix minor bug during user edit

* Missing changes + cleanup

* feat(wordpress): upgrade version, update plugins and themes

+++ srv/wordpress/wp-content/plugins/wp-graphql/src/Type/Object/MediaItemMeta.php
@@ -11,6 +11,8 @@ class MediaItemMeta {
 
 	/**
 	 * Register the MediaItemMeta Type
+	 *
+	 * @return void
 	 */
 	public static function register_type() {
 		register_graphql_object_type(
@@ -40,7 +42,7 @@ class MediaItemMeta {
 						'type' => 'String',
 					],
 					'focalLength'      => [
-						'type'    => 'Int',
+						'type'    => 'Float',
 						'resolve' => function( $meta, $args, $context, $info ) {
 							return ! empty( $meta['focal_length'] ) ? $meta['focal_length'] : null;
 						},
